core/flowfm-inspector/pyproject.toml
[tool.poetry]
name = "flowfm-inspector"
version = "0.1.0"
description = ""
authors = ["tegelaer <maarten.tegelaers@deltares.nl>"]

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^1.8.2"
fastapi = {extras = ["all"], version = "^0.70.1"}
hydrolib-core = "^0.2.0"
strawberry-graphql = {extras = ["fastapi"], version = "^0.93.10"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
